v2.0.0_2025.07.20-alpha.3
- 优化手表布局样式
- 优化视图圆角大小
- 优化输入框样式和消除按钮
- 优化数据存取逻辑
- 优化数据和缓存的清除逻辑
- 优化弹窗样式
- 优化输入框样式
- 优化背景样式
- 优化首页组件的显示样式
- 优化编辑显示方式的逻辑
- 优化手表端顶部按钮样式
- 取消手机端顶部按钮，改为浮动按钮
- 新增拖动排序的背景
- 屏蔽点击文本框的默认振动
- 更换未安装软件的图标

+++ app/src/main/java/cn/minimote/toolbox/helper/StoredToolHelper.kt
@@ -8,7 +8,8 @@ package cn.minimote.toolbox.helper
 
 import android.widget.Toast
 import cn.minimote.toolbox.R
-import cn.minimote.toolbox.constant.StoredActivity.STORED_FILE_NAME
+import cn.minimote.toolbox.constant.Config.ENCODING
+import cn.minimote.toolbox.constant.StoredTool.STORED_FILE_NAME
 import cn.minimote.toolbox.constant.Version
 import cn.minimote.toolbox.dataClass.StoredActivityContainer
 import cn.minimote.toolbox.dataClass.StoredTool
@@ -16,6 +17,7 @@ import cn.minimote.toolbox.viewModel.MyViewModel
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import java.io.File
+import java.io.FileOutputStream
 import java.lang.reflect.Type
 
 
@@ -24,7 +26,7 @@ object StoredToolHelper {
     private val gson = Gson()
 
     // 获取存储文件对象
-    private fun getStorageFile(viewModel: MyViewModel): File {
+    private fun getStoredFile(viewModel: MyViewModel): File {
         return File(viewModel.dataPath, STORED_FILE_NAME)
     }
 
@@ -34,28 +36,35 @@ object StoredToolHelper {
         viewModel: MyViewModel,
         storedToolList: MutableList<StoredTool>
     ) {
-        synchronized(this) {
-            val storedActivityContainer =
-                StoredActivityContainer(
-                    version = Version.STORED_ACTIVITY,
-                    data = storedToolList,
-                )
-            val json = gson.toJson(storedActivityContainer)
-            val file = getStorageFile(viewModel)
-//            Log.i("ActivityStorageHelper", "保存文件: ${file.absolutePath}")
-            file.writeText(json)
+        if(storedToolList.isNotEmpty()) {
+            synchronized(this) {
+                val storedActivityContainer =
+                    StoredActivityContainer(
+                        version = Version.STORED_ACTIVITY,
+                        data = storedToolList,
+                    )
+                val json = gson.toJson(storedActivityContainer)
+                val file = getStoredFile(viewModel)
+                FileOutputStream(file).use { outputStream ->
+                    outputStream.write(json.toByteArray(ENCODING))
+                }
+            }
+        } else {
+            deleteStorageFile(viewModel)
         }
     }
 
 
     // 加载存储列表
     fun loadStoredToolList(viewModel: MyViewModel): MutableList<StoredTool> {
-        val file = getStorageFile(viewModel)
-//        Log.i("ActivityStorageHelper", "读取文件: ${file.absolutePath}")
+        val file = getStoredFile(viewModel)
+        LogHelper.e("loadStoredToolList", "读取文件: ${file.absolutePath}")
 
         if(!file.exists()) {
+            LogHelper.e("loadStoredToolList", "文件不存在: ${file.absolutePath}")
             return mutableListOf()
         }
+        LogHelper.e("loadStoredToolList", "文件存在: ${file.absolutePath}")
 
         return synchronized(this) {
             try {
@@ -63,7 +72,7 @@ object StoredToolHelper {
                 if(json.isBlank()) {
                     // 文件存在但内容为空，视为无效文件
 //                    Log.e("ActivityStorageHelper", "文件内容为空")
-                    deleteFile(viewModel)
+                    deleteStorageFile(viewModel)
                     return mutableListOf()
                 }
 
@@ -72,14 +81,14 @@ object StoredToolHelper {
 
                 // 如果列表为空或者版本不一致，则删除文件
                 if(storedActivityContainer.data.isEmpty() || storedActivityContainer.version != Version.STORED_ACTIVITY) {
-                    deleteFile(viewModel)
+                    deleteStorageFile(viewModel)
                     return mutableListOf()
                 }
                 storedActivityContainer.data
             } catch(e: Exception) {
 //                Log.e("ActivityStorageHelper", "读取文件出错: ${e.message}", e)
                 e.printStackTrace()
-                deleteFile(viewModel)
+                deleteStorageFile(viewModel=viewModel, silence = false)
                 mutableListOf()
             }
         }
@@ -87,14 +96,20 @@ object StoredToolHelper {
 
 
     // 删除有问题的文件
-    private fun deleteFile(viewModel: MyViewModel) {
+    private fun deleteStorageFile(
+        viewModel: MyViewModel,
+        silence: Boolean = true, // 是否静默
+    ) {
         val context = viewModel.myContext
         // Log.e("ActivityStorageHelper", "删除文件")
-        Toast.makeText(
-            context, context.getString(R.string.data_error_and_delete_file),
-            Toast.LENGTH_LONG,
-        ).show()
-        val file = getStorageFile(viewModel)
+        if(!silence) {
+            Toast.makeText(
+                context, context.getString(R.string.data_error_and_delete_file),
+                Toast.LENGTH_LONG,
+            ).show()
+            return
+        }
+        val file = getStoredFile(viewModel)
         if(file.exists()) {
             file.delete()
         }
